% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_pred}
\alias{compare_pred}
\alias{compare_pred.saeTrafo}
\title{Compare predictions of model objects}
\usage{
compare_pred(object1, object2, MSE = FALSE, ...)

\method{compare_pred}{saeTrafo}(object1, object2, MSE = FALSE, ...)
}
\arguments{
\item{object1}{an object of type "saeTrafo".}

\item{object2}{an object of type "saeTrafo" or "emdi"
(\code{\link[emdi]{emdiObject}}).}

\item{MSE}{if \code{TRUE}, MSE estimates are returned. Defaults to
\code{FALSE} and than point estimates are returned.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Data frame containing the point estimates or the MSE estimates
(if \code{MSE} is set to \code{TRUE}\) of both objects. If column names are
duplicated, the suffixes "_1" and "_2" are added to their names. "_1" and
"_2" standing for object1 and object2, respectively.
}
\description{
Function \code{compare_pred} is a generic function used to compare
predictions of two model objects.

Method \code{compare_pred.saeTrafo} compares predictions of two saeTrafo
objects or a saeTrafo object and a emdi object.
}
\seealso{
\code{\link{emdi}}, \code{\link{NER_Trafo}},
\code{\link{saeTrafoObject}}
}
