% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_methods.R
\name{fixef}
\alias{fixef}
\alias{fixef.NER}
\alias{fixed.effects}
\alias{fixed.effects.NER}
\title{Extract fixed effects from an saeTrafo object}
\usage{
\method{fixef}{NER}(object, ...)

\method{fixed.effects}{NER}(object, ...)
}
\arguments{
\item{object}{an object of type "NER".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
A vector containing the fixed effects is returned.
}
\description{
Methods \code{fixef.NER} extract the fixed effects from an saeTrafo object of
class "NER".
}
\details{
The alias \code{fixed.effects} can also be used instead of
\code{fixef}. The generic function \code{fixef} is imported from package
\pkg{nlme} and re-exported to make the S3-methods available, even though the
\pkg{nlme} package itself is not loaded or attached. For default documentation,
see \code{\link[nlme]{fixed.effects}}.
}
\seealso{
\code{\link{NER_Trafo}}, \code{\link[nlme]{fixed.effects}}
}
