% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_methods.R
\name{getData}
\alias{getData}
\alias{getData.NER}
\title{Extract saeTrafo object data}
\usage{
\method{getData}{NER}(object, ...)
}
\arguments{
\item{object}{an object of type "NER".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
Data frame used to fit the model. For "NER" the (untransformed)
sample data is returned.
}
\description{
Methods \code{getData.NER} extract the data frame used to fit the model.
}
\details{
The generic function \code{getData} is imported from package nlme and
re-exported to make the S3-methods available, even though the nlme package
itself is not loaded or attached. For default documentation,
see \code{\link[nlme]{getData}}.
}
\seealso{
\code{\link{NER_Trafo}}, \code{\link[nlme]{getData}}
}
