% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_methods.R
\name{getGroups}
\alias{getGroups}
\alias{getGroups.NER}
\title{Extract grouping factors from an saeTrafo object}
\usage{
\method{getGroups}{NER}(object, ...)
}
\arguments{
\item{object}{an object of type "NER".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
A vector containing the grouping factors.
}
\description{
Method \code{getGroups.NER} extract grouping factors from a saeTrafo object.
}
\details{
The generic function \code{getGroups} is imported from package nlme and
re-exported to make the S3-methods available, even though the nlme package
itself is not loaded or attached. For default documentation,
see \code{\link[nlme]{getGroups}}.
}
\seealso{
\code{\link{NER_Trafo}}, \code{\link[nlme]{getGroups}}
}
