% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_methods.R
\name{getGroupsFormula}
\alias{getGroupsFormula}
\alias{getGroupsFormula.NER}
\title{Extract grouping formula from a saeTrafo object}
\usage{
\method{getGroupsFormula}{NER}(object, ...)
}
\arguments{
\item{object}{an object of type "NER".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
A one-sided formula.
}
\description{
Methods \code{getGroupsFormula.NER} extract the grouping formula from an
saeTrafo object.
}
\details{
The generic function \code{getGroupsFormula} is imported from package
nlme and re-exported to make the S3-methods available, even though the nlme package
itself is not loaded or attached. For default documentation,
see \code{\link[nlme]{getGroupsFormula}}.
}
\seealso{
\code{\link{NER_Trafo}} \code{\link[nlme]{getGroupsFormula}}
}
