% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_methods.R
\name{getResponse}
\alias{getResponse}
\alias{getResponse.NER}
\title{Extract response variable from an saeTrafo object}
\usage{
\method{getResponse}{NER}(object, ...)
}
\arguments{
\item{object}{an object of type "NER".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
Vector containing the response variable.
}
\description{
Methods \code{getResponse.NER} extract the response variable from a saeTrafo
object.
}
\details{
The generic function \code{getResponse} is imported from package
nlme and re-exported to make the S3-methods available, even though the nlme
package itself is not loaded or attached. For default documentation,
see \code{\link[nlme]{getResponse}}.
}
\seealso{
\code{\link{NER_Trafo}}, \code{\link[nlme]{getResponse}}
}
