% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.saeTrafo.R
\name{map_plot}
\alias{map_plot}
\title{Visualizes regional disaggregated estimates on a map}
\usage{
map_plot(
  object,
  MSE = FALSE,
  CV = FALSE,
  map_obj = NULL,
  map_dom_id = NULL,
  map_tab = NULL,
  color = c("white", "red4"),
  scale_points = NULL,
  guide = "colourbar",
  return_data = FALSE
)
}
\arguments{
\item{object}{an object of type \code{saeTrafo}, containing the estimates to
be visualized.}

\item{MSE}{optional logical. If \code{TRUE}, the MSE is also visualized.
Defaults to \code{FALSE}.}

\item{CV}{optional logical. If \code{TRUE}, the CV is also visualized.
Defaults to \code{FALSE}.}

\item{map_obj}{an \code{SpatialPolygonsDataFrame} object as defined by the
\pkg{sp} package on which the data should be visualized.}

\item{map_dom_id}{a character string containing the name of a variable in
\code{map_obj} that indicates the domains.}

\item{map_tab}{a \code{data.frame} object with two columns that match the
domain variable from the census data set (first column) with the domain
variable in the map_obj (second column). This should only be used if the IDs
in both objects differ.}

\item{color}{a \code{vector} of length 2 defining the lowest and highest
color in the plots.}

\item{scale_points}{a structure defining the lowest, the mid and the highest
value of the colorscale. If a numeric vector of length two is given, this
scale will be used for every plot. Alternatively, a list defining colors for
each plot separately may be given.}

\item{guide}{character passed to \code{scale_fill_gradient} from
\pkg{ggplot2}. Possible values are "none", "colourbar", and "legend".}

\item{return_data}{if set to \code{TRUE}, a fortified data frame including
the map data as well as the chosen indicators is returned. Customized maps
can easily be obtained from this data frame via the package \pkg{ggplot2}.
Defaults to \code{FALSE}.}
}
\value{
Creates the plots demanded, and, if selected, a fortified data.frame
containing the mapdata and chosen indicators.
}
\description{
Function \code{map_plot} creates spatial visualizations of the estimates
obtained by small area estimation methods.
}
\seealso{
\code{\link[maptools]{readShapePoly}}, \code{\link{sp}},
\code{\link{NER_Trafo}}, \code{\link{saeTrafoObject}}
}
