% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_area_size.R
\docType{data}
\name{pop_area_size}
\alias{pop_area_size}
\title{Aggregates from simulated eusilc population data: domain sizes}
\format{
A named numeric vector containing the number of individuals within
each domain. This numeric vector is named with the domain names.
}
\usage{
pop_area_size
}
\description{
This data contains aggregates from \code{\link{eusilcA_pop}}
which is based on \code{\link[simFrame]{eusilcP}} from package
\pkg{simFrame}.
}
\keyword{datasets}
