% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_cov.R
\docType{data}
\name{pop_cov}
\alias{pop_cov}
\title{Aggregates from simulated eusilc population data: domain-specific covariances}
\format{
A named list. Each element of the list contains the
domain-specific covariance matrice for p covariates for a specicfic domain.
The list is named with the respective domain name. The matrix within the list
has row and column names with the respective covariate names. The covariates
right of the ~ operator in \code{fixed} need to comprise.
}
\usage{
pop_cov
}
\description{
This data contains aggregates from \code{\link{eusilcA_pop}}
which is based on \code{\link[simFrame]{eusilcP}} from package
\pkg{simFrame}.
}
\keyword{datasets}
