% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_methods.R
\name{predict.NER}
\alias{predict.NER}
\title{Predictions from saeTrafo objects}
\usage{
\method{predict}{NER}(object, ...)
}
\arguments{
\item{object}{an object of type "saeTrafo".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
Data frame with domain predictors.
}
\description{
Method \code{predict.NER} extracts the direct estimates, the empirical
best linear unbiased or empirical best predictors for all domains from an
saeTrafo object.
}
