% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqnorm.R
\name{qqnorm.saeTrafo}
\alias{qqnorm.saeTrafo}
\alias{qqnorm.NER}
\title{Quantile-quantile plots for an saeTrafo object}
\usage{
\method{qqnorm}{saeTrafo}(y, color = c("blue", "lightblue3"), gg_theme = NULL, ...)

\method{qqnorm}{NER}(y, color = c("blue", "lightblue3"), gg_theme = NULL, ...)
}
\arguments{
\item{y}{a model object of type "saeTrafo" (see also \code{\link{NER_Trafo}}).}

\item{color}{a character vector with two elements. The first element defines
the color for the line in the QQ-plots, for the Cook's Distance plot and for
the Box-Cox plot. The second element defines the color for the densities.}

\item{gg_theme}{\code{\link[ggplot2]{theme}} list from package \pkg{ggplot2}.
For using this argument, package \pkg{ggplot2} must be loaded via
\code{library(ggplot2)}. See also Example 4.}

\item{...}{optional arguments passed to generic function.}
}
\value{
Two Q-Q plots in one grid obtained by \code{\link[ggplot2]{ggplot}}.
}
\description{
Normal quantile-quantile plots of the underlying model (see also
\code{\link{NER_Trafo}}) are obtained. The plots are obtained by
\code{\link[ggplot2]{ggplot}}.
}
\seealso{
\code{\link{saeTrafoObject}}, \code{\link{NER_Trafo}}
}
