% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_methods.R
\name{ranef}
\alias{ranef}
\alias{ranef.NER}
\alias{random.effects}
\alias{random.effects.NER}
\title{Extract random effects of saeTrafo object}
\usage{
\method{ranef}{NER}(object, ...)

\method{random.effects}{NER}(object, ...)
}
\arguments{
\item{object}{an object of type "NER".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
A vector containing the estimated random effects at domain level is
returned.
}
\description{
Method \code{ranef.NER} extract the fixed effects from an saeTrafo object.
}
\details{
The alias \code{random.effects} can also be used instead of
\code{ranef}. The generic function \code{ranef} is imported from package
\pkg{nlme} and re-exported to make the S3-methods available, even though the
\pkg{nlme} package itself is not loaded or attached. For default
documentation, see \code{\link[nlme]{random.effects}}.
}
\seealso{
\code{\link{NER_Trafo}}, \code{\link[nlme]{random.effects}}
}
