% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeTrafo.R
\docType{package}
\name{saeTrafo}
\alias{saeTrafo}
\title{A package for estimating unit-level small area models with transformed
dependent variable.}
\description{
The package \pkg{saeTrafo} supports estimating regional means ....
bias correction \cite{Molina and Martin (2019)}
}
\details{
The three estimation functions is called \code{\link{NER_Trafo}}.
For all functions, several methods are avalaible ....
}
\references{
Battese, G.E., Harter, R.M. and Fuller, W.A. (1988). An Error-Components
Model for Predictions of County Crop Areas Using Survey and Satellite Data.
Journal of the American Statistical Association, Vol.83, No. 401, 28-36. \cr \cr
}
